requirements.txt
openai==1.66.3
langchain-google-genai
google-generativeai
langchain
langchain-core
langchain-community
langchain-qdrant
qdrant-client
sentence-transformers
streamlit
audio-recorder-streamlit
python-dotenv